Identify measurable taxa

Identfies taxa above certain abundance and prevalence thresholds, that can be quantitatively analyzed

measurable_taxa(...)

# S3 method for default
measurable_taxa(Tab, min_reads_otu = 25,
  min_samples_otu = 5, method = "absolute", table = TRUE)

# S3 method for Dataset
measurable_taxa(Dat, min_reads_otu = 25,
  min_samples_otu = 5, method = "absolute", table = TRUE, clean = TRUE)

Arguments

Tab
Count table matrix. Samples as columns and OTUs as rows.

method
Either "absolute" (min absolute number of reads in min number of sample) or "AbsProp" (absolute number if reads and proportion of samples).

table
logical indicating whether an abundance table should be returned

Dat
A Dataset object.

clean
logical, indicating whether the function clean should be applied to the result.

Value

When table = FALSE, it returns a logical index vector that shows which taxa pass the thresholds.

When table = TRUE, it returns a matrix or Dataset object (depending on the method used), where samples that do not pass the thresholds have been removed

Details

Implments abundance filters defined in Benson et al.(2010) & Lundberg et al. (2012). It is recommended to threshold any Dataset before any quantitative analysis.

References

1. Benson AK, Kelly S a, Legge R, Ma F, Low SJ, Kim J, et al. Individuality in gut microbiota composition is a complex polygenic trait shaped by multiple environmental and host genetic factors. Proc Natl Acad Sci U S A. 2010 Nov 2;107(44):18933–8. Available from: http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=2973891&tool=pmcentrez&rendertype=abstract

2. Lundberg DS, Lebeis SL, Herrera Paredes S, Yourstone S, Gehring J, Malfatti S, et al. Defining the core Arabidopsis thaliana root microbiome. Nature. Nature Publishing Group; 2012 Aug 1;488(7409):86–90. Available from: http://www.nature.com/doifinder/10.1038/nature11237